+++ neural_network.h
@@ -11,14 +11,14 @@
  * Allocates empty model with weights and biases set to 0.
  */
 struct neural_network *create_model(size_t number_of_layers, size_t layers_size[],
-                                    size_t input_size, float (**activation)(float, int));
+                                    size_t input_size, double (**activation)(double, int));
 
 void free_neural_network(struct neural_network *nn);
 
 /*
  * Sets random weights to neural network. Weights are randomly chosen from a normal distribution of mean mu and stddev sigma.
  */
-void randomize_weights(struct neural_network *nn, float mu, float sigma, int use_bias);
+void randomize_weights(struct neural_network *nn, double mu, double sigma, int use_bias);
 
 /*
  * @param nn
@@ -26,7 +26,7 @@ void randomize_weights(struct neural_network *nn, float mu, float sigma, int use
  *
  * @return prediction vector
  */
-float *predict(struct neural_network *nn, float inputs[], size_t nb_inputs);
+double *predict(struct neural_network *nn, double inputs[], size_t nb_inputs);
 
 /*
  * Computes back propagation after a feed forward
@@ -36,8 +36,8 @@ float *predict(struct neural_network *nn, float inputs[], size_t nb_inputs);
  * @param learning_rate alpha, step size of gradient.
  * @param gamma momentum constant. If gamma = 0, doesn't have impact on computation. Lower gamma = lower momentum.
  */
-void back_propagate(struct neural_network *nn, float *output,
-                    float inputs[], float learning_rate, float gamma);
+void back_propagate(struct neural_network *nn, double *output,
+                    double inputs[], double learning_rate, double gamma);
 
 /*
  * Fits the model to given inputs and outputs.
@@ -50,8 +50,8 @@ void back_propagate(struct neural_network *nn, float *output,
  * @param learning_rate step size of gradient
  * @param gamma momentum constant. If gamma = 0, doesn't have impact on computation. Lower gamma = lower momentum.
  */
-void fit(struct neural_network *nn, size_t data_size, float *inputs[],
-         float *outputs[], size_t epochs, size_t batch_size, float learning_rate, float gamma);
+void fit(struct neural_network *nn, size_t data_size, double *inputs[],
+         double *outputs[], size_t epochs, size_t batch_size, double learning_rate, double gamma);
 
 /*
  * Evaluates a trained model over a new data set.
@@ -62,7 +62,7 @@ void fit(struct neural_network *nn, size_t data_size, float *inputs[],
  * @param loss loss function
  * @param verbose display mode. Set to true for prints.
  */
-float evaluate(struct neural_network *nn, size_t data_size,
-               float *inputs[], float *outputs[], float (*loss)(float *, float *, size_t), int verbose);
+double evaluate(struct neural_network *nn, size_t data_size,
+               double *inputs[], double *outputs[], double (*loss)(double *, double *, size_t), int verbose);
 
 #endif // __NEURAL_NETWORK_H__
